+++ NEWS.md
@@ -1,5 +1,4 @@
 # EDBdata 0.1.0
 
-# EDBdata 0.0.0.9000
-
+* Finished processing HAB satellite data for Franks Tract and Mildred Island which is in the `sat_ci_count_fr_mil` data object. This is the data used in the November draft of the EDB report.
 * Added a `NEWS.md` file to track changes to the package.