+++ NEWS.md
@@ -1,3 +1,5 @@
+# EDBdata (development version)
+
 # EDBdata 0.1.0
 
 * Finished processing HAB satellite data for Franks Tract and Mildred Island which is in the `sat_ci_count_fr_mil` data object. This is the data used in the November draft of the EDB report.